refactor: simplify Container import and export structure
